requirements.txt
autopep8
dhooks
fake-useragent
pyyaml
requests